Refactor Meal class
This commit applies the following changes:
  * Rename the AJAXRequests class to AJAXHelper
  * Extract AJAX requests out of the Meal class and into the AJAXHelper class
  * Extract logic for changing the HTML markup out of the Meal class and into the HTMLHelper class
