+++ lib/helpers/ajaxHelper.js
@@ -1,8 +1,9 @@
 const $ = require('jquery');
 const setUrl = require('../config/apiUrl').setUrl;
+const HTMLHelper = require('../helpers/htmlHelper');
 
 class AjaxRequests {
-    static get(){
+    static getFoods() {
         return $.ajax({
             type: 'GET',
             url: setUrl + '/api/v1/foods'
@@ -23,6 +24,21 @@ class AjaxRequests {
             url: `${setUrl}api/v1/foods/${id}`,
             data: { food: { calories: foodCalorie } }
         })
+    }
+
+    static postFoodMealJoinToServer(foodId, mealId) {
+      $.ajax({
+        method: 'POST',
+        url: `${setUrl}/api/v1/meals/${mealId}/foods/${foodId}`
+      });
+    }
+
+    static getMeals() {
+      return $.get(setUrl + '/api/v1/meals');
+    }
+
+    static getFoodMealRelationships(mealId) {
+      return $.get(`${setUrl}api/v1/meals/${mealId}/foods`);
     }
 }
