+++ lib/helpers/htmlHelper.js
@@ -1,4 +1,5 @@
 const $ = require('jquery');
+const AjaxRequests = require('./ajaxHelper');
 
 class HTMLHelper {
     static toHtml(args){
@@ -28,6 +29,31 @@ class HTMLHelper {
 
     static resetCheckbox(){
         $('input:checked').prop('checked', false);
+    }
+
+    static setMealIdsInHTML(meals) {
+      meals.forEach((meal) => {
+        const mealId = meal.id;
+        const mealName = meal.name.toLowerCase();
+        $(`.${mealName}`).attr('data-id', mealId);
+      })
+    }
+
+    static insertFoodMealRelationsInHTML() {
+      const mealNames = ['breakfast', 'lunch', 'dinner', 'snack'];
+      mealNames.forEach((mealName) => {
+        const mealId = $(`.${mealName}`)[0].dataset.id;
+        AjaxRequests.getFoodMealRelationships(mealId)
+        .then((meal) => {
+          meal.foods.forEach((food) => {
+            const foodName = food.name;
+            const foodCalorie = food.calories;
+            const foodId = food.id;
+            const foodRowTemplate = HTMLHelper.toHtml({foodName: foodName, foodCalorie: foodCalorie, id: foodId, checkbox: false});
+            $(`.${mealName} tr:first`).after(foodRowTemplate);
+          });
+        });
+      });
     }
 }
