+++ lib/models/food.js
@@ -6,7 +6,7 @@ const AjaxRequests = require('../helpers/ajaxHelper');
 class Food {
 
     static all() {
-        AjaxRequests.get()
+        AjaxRequests.getFoods()
         .then( (response) => {
             const use_checkbox = document.title === "Quantified Self - Manage Foods" ? false : true;
             response.forEach( (food) => {
