+++ lib/models/meal.js
@@ -1,6 +1,7 @@
 const $ = require('jquery');
 const HTMLHelper = require('../helpers/htmlHelper');
 const setUrl = require('../config/apiUrl').setUrl;
+const AjaxRequests = require('../helpers/ajaxHelper');
 
 class Meal {
   static addFoodsTo(mealName) {
@@ -18,7 +19,7 @@ class Meal {
           let mealId = $(`.${mealName}`)[0].dataset.id;
 
           // post the new food/meal relationship to the server
-          Meal.postFoodMealJoinToServer(foodId, mealId);
+          AjaxRequests.postFoodMealJoinToServer(foodId, mealId);
 
           // add the food to the meal which corresponds to the clicked meal button
           const foodRowTemplate = HTMLHelper.toHtml({foodName: foodName, foodCalorie: foodCalorie, id: foodId, checkbox: false});
@@ -27,13 +28,6 @@ class Meal {
       }
     });
 
-  }
-
-  static postFoodMealJoinToServer(foodId, mealId) {
-    $.ajax({
-      method: 'POST',
-      url: `${setUrl}/api/v1/meals/${mealId}/foods/${foodId}`
-    });
   }
 
   static handleAddFoodsToMeals() {
@@ -59,37 +53,12 @@ class Meal {
   }
 
   static populateMealPage() {
-    Meal.getMealIds()
-    .then(() => {
-      Meal.getFoodMealRelationships();
+    AjaxRequests.getMeals()
+    .then((meals) => {
+      HTMLHelper.setMealIdsInHTML(meals);
     })
-  }
-
-  // get meal ids from the server and set them as data attributes on the meal HTML tags
-  static getMealIds() {
-    return $.get(setUrl + '/api/v1/meals', (meals) => {
-      meals.forEach((meal) => {
-        const mealId = meal.id;
-        const mealName = meal.name.toLowerCase();
-        $(`.${mealName}`).attr('data-id', mealId);
-      });
-    });
-  }
-
-  static getFoodMealRelationships() {
-    const meals = ['breakfast', 'lunch', 'dinner', 'snack'];
-    meals.forEach((mealName) => {
-      const mealId = $(`.${mealName}`)[0].dataset.id;
-      $.get(`${setUrl}api/v1/meals/${mealId}/foods`)
-      .then((meal) => {
-        meal.foods.forEach((food) => {
-          const foodName = food.name;
-          const foodCalorie = food.calories;
-          const foodId = food.id;
-          const foodRowTemplate = HTMLHelper.toHtml({foodName: foodName, foodCalorie: foodCalorie, id: foodId, checkbox: false});
-          $(`.${mealName} tr:first`).after(foodRowTemplate);
-        });
-      });
+    .then(() => {
+      HTMLHelper.insertFoodMealRelationsInHTML();
     });
   }
 }
